Fix app structure
