+++ app-methods/author_actions.rb
@@ -1,15 +1,3 @@
-# def author_display
-#     if @authors.empty?
-#       puts "There isn't any author in our catalog."
-#       sleep(1)
-#     else
-#       @authors.each_with_index do |element, idx|
-#         line = "#{idx + 1}) Author: #{element.first_name} #{element.last_name}"
-#         puts line
-#       end
-#       sleep(2)
-#     end
-#   end
 module AuthorActions
   def author_display
     if @authors.empty?
