+++ app-methods/music_genre_module.rb
