+++ data-functions/author_json.rb
