+++ main.rb
@@ -2,26 +2,48 @@ require_relative './app'
 
 def list_of_options
   puts 'Please choose an option by entering a number:'
-  puts '1 list all music albums'
-  puts '2 create a new music album'
-  puts '3'
-  puts '4'
-  puts '5'
-  puts '6'
-  puts '7 - exit'
+  puts '[1] - List all books'
+  puts '[2] - List all music albums'
+  puts '[3] - List all games'
+  puts '[4] - List all genres'
+  puts '[5] - List all labels'
+  puts '[6] - List all authors'
+  puts '[7] - Add a book'
+  puts '[8] - Add a music album'
+  puts '[9] - Add a game'
+  puts '[0] - exit'
 end
 
-def actions(option, _app)
+# rubocop:disable Metrics/CyclomaticComplexity
+def actions(option, app)
   case option
-  when 1 then list_music_albums
-  when 2 then add_music_album
+  when 1
+    app.list_all_books
+  when 2
+    puts 'List all music albums'
+  when 3
+    app.game_display
+  when 4
+    puts 'List all genres'
+  when 5
+    app.list_labels
+  when 6
+    app.author_display
+  when 7
+    app.create_book
+  when 8
+    puts 'Add music album'
+  when 9
+    app.game_create
   end
 end
+# rubocop:enable Metrics/CyclomaticComplexity
 
 def main
   app = App.new
   puts 'Welcome to our app!'
   puts ''
+  app.load_data
 
   loop do
     list_of_options
@@ -33,11 +55,13 @@ def main
       puts ''
       sleep(1)
     end
-    break if option == 7
+    break if option.zero?
   end
+  # Save data Functions
+  app.save_data
 
   puts 'Thanks for using the app!'
-  save_albums(@albums)
 end
 
 main
+
