+++ methods/genre.rb
