+++ methods/music_album.rb
